+++ crate_anon/crateweb/specimen_secret_local_settings/crateweb_local_settings.py
@@ -36,12 +36,15 @@ https://crateanon.readthedocs.io/en/latest/website_config/web_config_file.html
 
 """
 
+import logging
 import os
 from typing import List, TYPE_CHECKING
 if TYPE_CHECKING:
     from django.http.request import HttpRequest
 
-raise Exception(
+log = logging.getLogger(__name__)
+
+log.critical(
     "Well done - CRATE has found your crate_local_settings.py file at {}. "
     "However, you need to configure it for your institution's set-up, and "
     "remove this line.".format(os.path.abspath(__file__)))
@@ -314,6 +317,10 @@ ARCHIVE_TEMPLATE_DIR = "/home/somewhere/my_archive_templates"
 ARCHIVE_STATIC_DIR = "/home/somewhere/my_archive_templates/static"
 ARCHIVE_ROOT_TEMPLATE = "root.mako"
 ARCHIVE_ATTACHMENT_DIR = "/home/somewhere/my_archive_attachments"
+ARCHIVE_CONTEXT = {}
+CACHE_CONTROL_MAX_AGE_ARCHIVE_ATTACHMENTS = 0
+CACHE_CONTROL_MAX_AGE_ARCHIVE_TEMPLATES = 0
+CACHE_CONTROL_MAX_AGE_ARCHIVE_STATIC = 0
 
 
 # =============================================================================