+++ docs/source/conf.py
@@ -262,7 +262,7 @@ autodoc_mock_imports = [
 
 # sys.path.insert(0, os.path.join(os.path.abspath('.'), '../../myproj'))
 
-# ... see crate_anon/crateweb/config/settings.py  # noqa
+# ... see crate_anon/crateweb/config/settings.py
 os.environ['DJANGO_SETTINGS_MODULE'] = 'crate_anon.crateweb.config.settings'
 django.setup()